+++ src/components/chat/ChatArea.tsx
@@ -2,8 +2,8 @@ import React, {useEffect, useRef, useState} from "react";
 import {Ionicons} from "@expo/vector-icons";
 import {Animated, Easing, ScrollView, Text, TextInput, TouchableOpacity, View,} from "react-native";
 import {Conversation} from "@/src/models/Conversation";
-import EmojiPicker from "./EmojiPicker";
-import StickerPicker from "./StickerPicker";
+import EmojiPicker from "./input/EmojiPicker";
+import StickerPicker from "./input/StickerPicker";
 import {Shadows} from "@/src/styles/Shadow";
 import {Message, MessageType} from "@/src/models/Message";
 import {MessageService} from "@/src/api/services/MessageService";
