+++ src/components/chat/input/StickerPicker.tsx
